See description
Fixed filepaths.
Added a field for the destination file to the structure.
Creating directories if they were deleted when being restored.
Opening destination files with O_TRUNC, need to do this is the source file is smaller so the destination file is emptied out first.
Opening destination file with 0777 so if this file is being created there are the correct permissions.
Added a function to create the directories so code wasn't repeated.